+++ packages/checkbox-group/test/checkbox-group.test.js
@@ -1,5 +1,5 @@
 import { expect } from '@esm-bundle/chai';
-import { fixtureSync, nextFrame, nextRender } from '@vaadin/testing-helpers';
+import { fixtureSync, nextFrame } from '@vaadin/testing-helpers';
 import { sendKeys } from '@web/test-runner-commands';
 import sinon from 'sinon';
 import '../vaadin-checkbox-group.js';
@@ -299,141 +299,6 @@ describe('vaadin-checkbox-group', () => {
       group.appendChild(checkbox);
       await nextFrame();
       expect(console.warn.called).to.be.false;
-    });
-  });
-
-  describe('initial validation', () => {
-    let validateSpy;
-
-    beforeEach(() => {
-      group = document.createElement('vaadin-checkbox-group');
-      validateSpy = sinon.spy(group, 'validate');
-    });
-
-    afterEach(() => {
-      group.remove();
-    });
-
-    it('should not validate by default', async () => {
-      document.body.appendChild(group);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-
-    it('should not validate when the field has an initial value', async () => {
-      group.value = ['en'];
-      document.body.appendChild(group);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-
-    it('should not validate when the field has an initial value and invalid', async () => {
-      group.value = ['en'];
-      group.invalid = true;
-      document.body.appendChild(group);
-      await nextRender();
-      expect(validateSpy.called).to.be.false;
-    });
-  });
-
-  describe('validation', () => {
-    beforeEach(async () => {
-      group = fixtureSync(`
-        <vaadin-checkbox-group>
-          <vaadin-checkbox name="language" value="en" label="English">/vaadin-checkbox>
-          <vaadin-checkbox name="language" value="fr" label="Français"></vaadin-checkbox>
-          <vaadin-checkbox name="language" value="de" label="Deutsch">/vaadin-checkbox>
-        </vaadin-checkbox-group>
-      `);
-      checkboxes = group.querySelectorAll('vaadin-checkbox');
-      await nextFrame();
-    });
-
-    it('should not have invalid attribute initially', () => {
-      expect(group.hasAttribute('invalid')).to.be.false;
-    });
-
-    it('should not add invalid attribute if required attribute is not present', () => {
-      checkboxes[0].checked = true;
-      checkboxes[0].checked = false;
-      expect(group.hasAttribute('invalid')).to.be.false;
-    });
-
-    it('should add invalid attribute if required attribute is present and checkbox group value is empty', () => {
-      group.required = true;
-      checkboxes[0].checked = true;
-      checkboxes[0].checked = false;
-      expect(group.hasAttribute('invalid')).to.be.true;
-    });
-
-    it('should remove invalid attribute if checkbox group value is not empty', () => {
-      group.required = true;
-      checkboxes[0].checked = true;
-      checkboxes[0].checked = false;
-      expect(group.hasAttribute('invalid')).to.be.true;
-
-      checkboxes[0].checked = true;
-      expect(group.hasAttribute('invalid')).to.be.false;
-    });
-
-    it('should run validation only once when adding a value', () => {
-      const spy = sinon.spy(group, 'validate');
-      group.value = ['en', 'fr'];
-      expect(spy.calledOnce).to.be.true;
-    });
-
-    it('should run validation only once when removing a value', () => {
-      const spy = sinon.spy(group, 'validate');
-      group.value = ['en', 'fr'];
-      spy.resetHistory();
-      group.value = ['en'];
-      expect(spy.calledOnce).to.be.true;
-    });
-
-    it('should run validation and set invalid when field is required and user blurs out of the group', async () => {
-      group.required = true;
-
-      // Focus on the first checkbox.
-      await sendKeys({ press: 'Tab' });
-      // Move focus out of the checkbox group.
-      await sendKeys({ down: 'Shift' });
-      await sendKeys({ press: 'Tab' });
-      await sendKeys({ up: 'Shift' });
-
-      expect(group.invalid).to.be.true;
-    });
-
-    it('should not run validation while user is tabbing between checkboxes inside of the group', async () => {
-      group.required = true;
-      const spy = sinon.spy(group, 'validate');
-
-      // Focus on the first checkbox.
-      await sendKeys({ press: 'Tab' });
-      // Focus on the second checkbox.
-      await sendKeys({ press: 'Tab' });
-
-      expect(spy.called).to.be.false;
-    });
-
-    it('should fire a validated event on validation success', () => {
-      const validatedSpy = sinon.spy();
-      group.addEventListener('validated', validatedSpy);
-      group.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.true;
-    });
-
-    it('should fire a validated event on validation failure', () => {
-      const validatedSpy = sinon.spy();
-      group.addEventListener('validated', validatedSpy);
-      group.required = true;
-      group.validate();
-
-      expect(validatedSpy.calledOnce).to.be.true;
-      const event = validatedSpy.firstCall.args[0];
-      expect(event.detail.valid).to.be.false;
     });
   });
